To be squashed - Separate facade/promise from facade/async to avoid importing rxjs in XHRImpl (and by extension, in benchmarks)
