+++ modules/angular2/src/core/compiler/xhr_impl.ts
@@ -1,5 +1,5 @@
 import {Injectable} from 'angular2/src/core/di';
-import {Promise, PromiseWrapper, PromiseCompleter} from 'angular2/src/core/facade/async';
+import {Promise, PromiseWrapper, PromiseCompleter} from 'angular2/src/core/facade/promise';
 import {isPresent} from 'angular2/src/core/facade/lang';
 import {XHR} from './xhr';
